requirements.txt
asgiref==3.8.1
Django==5.1.7
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
